Fixed a bug where adding a review would change the business's location

+++ WorkOut/app/src/main/java/Model/Location.java
@@ -39,8 +39,8 @@ public class Location implements Parcelable
 
     protected Location(Parcel in)
     {
-        longitude = in.readLong();
-        latitude = in.readLong();
+        longitude = in.readDouble();
+        latitude = in.readDouble();
     }
 
     @Override
